requirements.txt
flask
redis
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
Flask-Migrate
mysqlclient
pytest
webargs